+++ bin/setup/system-roms-overlays.sh
@@ -116,6 +116,8 @@ EOF
 
   # Remove old, unused emulator overlay configs
   while read library_name; do
+    [ ! -d "$retroarch_config_dir/config/$library_name" ] && continue
+
     while read path; do
       [ ! "${installed_files["$path"]}" ] && rm -v "$path"
     done < <(find "$retroarch_config_dir/config/$library_name" -name '*.cfg')
